requirements.txt
Django
djangorestframework
django-filter
Pillow